app/requirements.txt
bandit~=1.8.0
coverage~=7.10.0
pip-audit~=2.9.0
pre-commit~=4.3.0
pytest~=8.4.0
requests-mock~=1.12.0
ruff~=0.12.0
schemathesis~=4.1.0
starlette
pydantic
pydantic_core
typing_extensions
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.2
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
argon2-cffi